trying to add no-overlap to get_z_encoder_for
